refactor: rename oh-my-opencode files to oh-my-openagent via git mv
- git mv bin/oh-my-opencode.js -> bin/oh-my-openagent.js
- git mv src/config/schema/oh-my-opencode-config.ts -> oh-my-openagent-config.ts
- Update import path in src/config/schema.ts
- Update self-reference comment in bin file

Wave 2 Task 4 complete.
