+++ src/config/schema.ts
@@ -14,7 +14,7 @@ export * from "./schema/git-env-prefix"
 export * from "./schema/git-master"
 export * from "./schema/hooks"
 export * from "./schema/notification"
-export * from "./schema/oh-my-opencode-config"
+export * from "./schema/oh-my-openagent-config"
 export * from "./schema/ralph-loop"
 export * from "./schema/runtime-fallback"
 export * from "./schema/skills"
